+++ src/plot_recipes/plot_recipes.jl
@@ -5,7 +5,7 @@ using RecipesBase
 
     (x, y) = xx
 
-    alpha --> 0.5
+    seriesalpha --> 0.5
     seriestype := :shape
 
     x = [x.lo, x.hi, x.hi, x.lo]
@@ -32,7 +32,7 @@ end
 
     end
 
-    alpha --> 0.5
+    seriesalpha --> 0.5
 
     #x = xs
     #y = ys